linter issue fix autops

+++ features/top_ads/top_ads_management_common/src/main/java/com/tokopedia/topads/common/domain/model/GetVariantByIdResponse.kt
@@ -25,7 +25,7 @@ data class GetVariantByIdResponse(
 
 data class GetVariantByIdInput(
     @SerializedName("shopId")
-    val shopId: Int,
+    val shopId: Long,
     @SerializedName("clientId")
-    val clientId: Int
+    val clientId: Long
 ) {}
